requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
chromadb
pypdf
docx2txt
pandas
openpyxl
python-pptx
markdown
streamlit
fastapi
uvicorn
python-dotenv
tqdm
pydantic
tiktoken
python-multipart
unstructured
